+++ ext/geos_c_impl/geometry_collection.c
@@ -522,6 +522,11 @@ method_multi_polygon_centroid(VALUE self)
   RGeo_GeometryData* self_data;
   const GEOSGeometry* self_geom;
 
+  if (RGEO_FACTORY_DATA_PTR(self)->flags & RGEO_FACTORYFLAGS_SUPPORTS_Z_OR_M) {
+    rb_raise(rb_eRGeoUnsupportedOperation,
+             "Geos cannot compute centroid for more than 2 dimensions");
+  }
+
   result = Qnil;
   self_data = RGEO_GEOMETRY_DATA_PTR(self);
   self_geom = self_data->geom;