+++ History.md
@@ -7,6 +7,7 @@
 **Minor Changes**
 
 * Add `simplify_polygon_hull` method to the CAPI factory (@oleksii-leonov) [#366](https://github.com/rgeo/rgeo/pull/366)
+* Prevent `#centroid` to be called in more than 2d in CAPI.
 
 **Bug Fixes**